+++ package.json
@@ -15,11 +15,17 @@
   "author": "Brian Leathem",
   "license": "MIT",
   "dependencies": {
+    "bluebird": "^3.4.7",
     "lodash": "^4.7.0",
     "mediator-js": "^0.9.9",
-    "bluebird": "^3.4.7"
+    "ts-node": "^2.0.0"
   },
   "devDependencies": {
+    "@types/bluebird": "^3.0.37",
+    "@types/chai": "^3.4.34",
+    "@types/lodash": "^4.14.52",
+    "@types/mocha": "^2.2.38",
+    "@types/sinon": "^1.16.34",
     "chai": "^3.5.0",
     "grunt": "^1.0.1",
     "grunt-eslint": "^19.0.0",
@@ -27,6 +33,8 @@
     "load-grunt-tasks": "^3.5.2",
     "mocha": "^3.2.0",
     "nsp": "^2.6.2",
-    "sinon": "^1.17.7"
+    "sinon": "^1.17.7",
+    "ts-node": "^2.0.0",
+    "typescript": "^2.1.5"
   }
 }
