+++ typings.json
@@ -1,0 +1,6 @@
+{
+  "dependencies": {
+    "bluebird": "registry:npm/bluebird#3.4.1+20160909132857",
+    "lodash": "registry:npm/lodash#4.0.0+20161015015725"
+  }
+}
